Move to fuzz module
